[fix]  代码静态安全扫描问题整改



# message auto-generated for no-merge-commit merge:
!1304 merge fix/security-issues into master

[fix]  代码静态安全扫描问题整改

Created-by: Hututu-w
Commit-by: Hututu-w
Merged-by: chenyz6
Description: # 合入来源

- [ ] 需求
- [ ] 问题单
- [x] issue

# 问题/功能描述

代码静态安全扫描发现的若干缺陷，涉及三类问题：

1. **空指针风险**：`offload_group_pack_copy` 对外 API 入口未对地址参数判空；`SendMessageNonBlock` 未对 `AccDataBuffer::Create` 返回值判空。
2. **整数溢出风险**：`GroupAllGather` 中 `sendSize * size` 为 uint32 乘法可能回绕导致校验失效；`ImportRemoteMemKeysLocked` 中 `memDescLen` 为 uint32 加法可能回绕绕过边界检查；`TimeNs` 中 `timeValue * 1000` 中间值在长时间运行后可能超出 uint64 范围。
3. **资源泄漏风险**：`AllocSwapMemory`（URMA/RDMA）失败分支依赖未赋值的成员变量清理，导致已分配内存泄漏；`InitStreamNotifyBuf` 失败分支未释放已分配设备内存；`CreateConnectionToServers` 错误路径未回滚已创建 socket；`PrepareLaunchBuffer` 分配顺序不满足异常安全。

# 修改方案描述

- acc_offload: `offload_group_pack_copy` 入口对全部 6 个地址参数判空并记录日志
- smem_net_group_engine: `GroupAllGather` 两处 `sendSize * size` 乘法提升为 `uint64_t` 计算
- smem_tcp_config_store: `SendMessageNonBlock` 对 `AccDataBuffer::Create` 结果判空
- hybm_data_op_device_urma: `AllocSwapMemory` 的 AddVaInfo 失败分支直接 `AclrtFree(ptr)` 释放
- hybm_data_op_device_rdma: `AllocSwapMemory` 的 AddVaInfo 失败分支显式 `HalHostUnregisterEx` + `HalMemFree` 回收
- device_urma_transport_manager: `memDescLen` 类型由 `uint32_t` 改为 `uint64_t`，消除加法回绕
- joinable_ranks_qp_manager: `CreateConnectionToServers` 引入 rollback lambda，4 个错误出口统一回滚已创建 socket
- device_rdma_transport_manager: `InitStreamNotifyBuf` 的 AclrtMemcpy 失败分支补充 `AclrtFree(ptr)`
- device_kernel_helper: `PrepareLaunchBuffer` 调整分配顺序，先构造 hostBuf（RAII）再分配设备内存
- mf_monotonic_time: `TimeNs` 改为先除后乘 + 余数补偿，规避中间值溢出

# 是否涉及UT/ST

- [ ] 是
- [x] 否（说明理由）：均为缺陷修复，不改变正常路径的功能语义与对外接口行为，无新增功能分支需要覆盖

# 开发自检

- [x] 规范：未超过1k；超过1k已完成备案；反合入需要标题注明【反合】
- [x] 规范：已关联issue或里程碑
- [x] 规范：不需要UT/ST且已说明理由；需要且已包含UT/ST
- [x] 安全：涉及对外接口/参数/环境变量改动，并同步修改资料（未改动对外接口定义，仅增加入参校验，无资料影响）
- [x] 安全：未引入三方开源软件；引入三方开源软件并经过安全评审
- [x] 安全：未新增对外接口/参数；新增对外接口/参数有校验（本 MR 为既有接口补充入参校验）
- [x] 规范：未涉及资料改动；涉及资料改动并已完成修改

See merge request: Ascend/memfabric_hybrid!1304

+++ src/hybm/csrc/transport/protocol/device/urma/device_urma_transport_manager.cpp
@@ -1405,7 +1405,7 @@ Result DeviceUrmaTransportManager::ImportRemoteMemKeysLocked(uint32_t peerRank, 
 
         const uint64_t remoteSize = exportDesc.size;
         const uint64_t memTag = exportDesc.memTag;
-        const uint32_t memDescLen = sizeof(UrmaExportDesc) + exportDesc.hcommDescLen;
+        const uint64_t memDescLen = sizeof(UrmaExportDesc) + exportDesc.hcommDescLen;
 
         // Validate total payload fits within the key data area
         if (memDescLen + exportDesc.devTransFlagDescLen > DEVICE_URMA_EXPORT_KEY_DATA_BYTES) {
